# docs/swagger.yaml
definitions:
  utils.ErrorRespond:
    properties:
      code:
        type: integer
      data:
        type: string
      status:
        type: string
    type: object
  utils.SuccessRespond:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
  utils.ValidationErrorRespond:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
info:
  contact:
    email: hello@aasumitro.id
    name: '@aasumitro'
    url: https://aasumitro.id/
  license:
    name: MIT
    url: https://github.com/aasumitro/karlota/blob/master/LICENSE
paths:
...
  /v1/update/fcm:
    post:
      consumes:
      - multipart/form-data
      description: Generate Access Token (JWT).
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: firebase cloun messaging token
        in: formData
        name: fcm_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: CREATED_RESPOND
          schema:
            $ref: '#/definitions/utils.SuccessRespond'
        "400":
          description: BAD_REQUEST_RESPOND
          schema:
            $ref: '#/definitions/utils.ErrorRespond'
        "422":
          description: UNPROCESSABLE_ENTITY_RESPOND
          schema:
            $ref: '#/definitions/utils.ValidationErrorRespond'
        "500":
          description: INTERNAL_SERVER_ERROR_RESPOND
          schema:
            $ref: '#/definitions/utils.ErrorRespond'
      summary: Update FCM TOKEN
      tags:
      - AccountHandler
...
swagger: "2.0"